pyproject.toml
[project]
name = "projectsentinelpsca"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-community>=0.4.1",
    "langchain-core>=1.2.18",
    "langchain-groq>=1.1.2",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.1.0",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "tavily-python>=0.7.23",
]
